Updated changes from slackware sysvinit-scripts
In rc.M, don't run rc.wicd if /usr/sbin/wicd does not exist.  This
change prevents a leftover rc.wicd from blocking NetworkManager startup.

Changed the path in rc.M for wicd from relative to absolute.  It works
either way, but it should be an absolute path just to be consistent with
the rest of the script.